git-svn-id: https://plugins.svn.wordpress.org/sitepress-multilingual-cms/branches/new-translation-workflow2@313701 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ inc/sitepress-schema.php
@@ -291,9 +291,12 @@ function icl_sitepress_activate(){
    if(get_option('icl_sitepress_version')){
        icl_plugin_upgrade();               
    }
-                  
-   delete_option('icl_sitepress_version');   
-   add_option('icl_sitepress_version', ICL_SITEPRESS_VERSION, '', true);
+   
+   // don't set the new version if a multi-step upgrade is in progress   
+   if(!defined('ICL_MULTI_STEP_UPGRADE')){
+       delete_option('icl_sitepress_version');   
+       add_option('icl_sitepress_version', ICL_SITEPRESS_VERSION, '', true);
+   }               
 
     
     $iclsettings = get_option('icl_sitepress_settings');
